chore: move onboarding component
